Finish Wonder Trade
This is a basic Wonder Trade functionality that works with little added fluff. The system chooses from all base form non-legendary Pokemon and evolves them if appropriate. This is to avoid biasing the trades towards bigger evolution lines; each line has an equal chance of appearing. OT Names are randomly chosen from a list of 500 of the most common US names from 2005, with names longer than seven characters from the source list skipped.

CreateWonderTrade is the special that needs to be called in a script to initiate a Wonder Trade. The script should be set up just like an ingame trade, but the trade identifier should be 0xFF instead of one of the ingame trade slots.