Framework for Egg Sprite Table & Show on Summary Screen
Code for displaying unique Egg sprites for each Pokemon capable of being hatched. Uses a new sprite table (egg_pic_table.inc). Graphics loading functions have been modified to check for Eggs and load the appropriate Egg sprite from the new table. The palette is taken from the Pokemon inside the Egg.

To add new Egg sprites, put them in the graphics/eggs/POKEMONNAME folder. They should use the same palette as the front pic of the Pokemon and be named "pic.png". Then, update the aforementioned sprite table so that the Pokemon's entry uses obj_tiles gMonEggPic_ instead of the placeholder obj_tiles gMonFrontPic_. Then, put your new sprite in graphics.inc, labelling it "gMonEggPic_POKEMONNAME".

Still TODO:
Show sprites in PC screen instead of generic Egg sprite.

Modify egg_hatch.c so that the hatching screen uses the unique sprites.

Handle Egg menu icons in both the Summary Screen and PC.

Create and include the remainder of the unique Egg sprites. Only ones in this commit are Treecko, Torchic, Mudkip, and Togepi. No icons are done yet.

Set Up Egg Icons and Make Trading Show Eggs

Unique Egg icons are set up with this commit and work in the party and trade screens. Egg icons must share the same palette as the species' icon, be put in the graphics/eggs/POKEMONNAME folder, and be named icon.png. They need to be added to graphics.h and graphics.inc, and added to gMonEggIconTable in pokemon_icon.c as well before they will work.

Work done to trade.c to show the icons in the trade menu as well as show the appropriate Egg sprites in the actual trade scene.

Still TODO:
Get the Egg icons and front sprites working in the Pokemon PC.

Finish the remaining Egg sprites and icons. Only Treecko, Torchic, Mudkip, and Togepi are included in this commit.

Make the hatching scene display the correct Egg sprite.

Track down remaining cases where an Egg might be shown and ensure that they display correctly.

Get Unique Eggs Working in PC

This commit modifies the Pokemon Storage System so that it properly displays the Unique Eggs.

Get Unique Eggs Working in Hall of Fame

This commit gets the unique Eggs showing properly in the Hall of Fame.

The code now checks if a Pokemon's species2 field shows it is an Egg. When it sees this, it sets the HoF data's species field to the baby's species and the HoF data's level field to 0. Now the checks for loading the Pokemon's sprite and printing its info check to see if the level is 0 and treat the entry as an Egg if so.

No changes were made to the existing HoF struct to do this, but HoF entries with Eggs from before this change or from a vanilla Ruby save will display incorrectly, as they did not set the level of Eggs to 0.

Make Egg Hatching Scene Show Unique Sprites

This commit makes it so that the Egg hatching screen shows the correct sprites. The implementation is not ideal, but it is functional and can be revisited later.

For this to work, unique versions of egg_hatch.png and egg_shard.png (found in graphics/misc) have to be created for each hatchable Pokemon. They should use the same palette as that Pokemon's front sprites, and be saved in with the other Egg graphics.

To get them to work, declare the sprites at the top of egg_hatch.c with the others (search //Unique Eggs and follow the format). Then, include the new Pokemon entry into the switch statement in the determineHatchGraphics function by following the format. This commit includes the hatching scene graphics for the four Pokemon that already have Egg graphics (Togepi, Treecko, Torchic, Mudkip).

I believe that this is the last place where unique Eggs need to be considered. This feature is functionally complete if so.

Still TODO:
Finish the remainder of the Egg sprites. There are 180 Pokemon that can be found in Eggs. Each needs a front sprite, a two-frame icon, cracking graphics for the hatching screen, and egg shard graphics for the hatching screen.

Add New Egg Sprites

This commit adds Egg Sprites for Vulpix, Oddish, Slowpoke, Sentret, and Aipom.

Eggs finished: 9
Eggs to go: 171
Eggs total :180

How to Properly Add Eggs to Game:
(These instructions assume you're following the existing naming scheme for the files.)

[*]Ensure all the images are in the appropriate folder (graphics/eggs/POKEMONNAME). They should be named pic, icon, egg_hatch, and egg_shard.

[*]Add the icon and pic files to the end of graphics.inc.

[*]Add the icon to graphics.h. The Egg icon entries start at line 2470.

[*]Edit the Pokemon's entry in the gMonEggIconTable of pokemon_icon.c. Change the name of the entry from gMonIcon_POKEMONNAME to gMonEggIcon_POKEMONNAME. The table starts on line 463.

[*]Edit the Pokemon's entry in the gMonEggPicTable of egg_pic_table.inc. Change the name of the entry from obj_tiles gMonFrontPic_POKEMONNAME to obj_tiles gMonEggPic_POKEMONNAME.

[*]At the top of egg_hatch.c, add your egg_hatch and egg_shard graphics to the other declarations. They start on line 85.

[*]In egg_hatch.c, add an entry for the species to the switch statement in the determineHatchGraphics() function, following the format of the other entries.

Add 6 More Eggs

This commit adds Egg sprites for Bellsprout, Eevee, Cyndaquil, Ledyba, Phanpy, and Larvitar. It also tweaks Sentret's Egg a little.

Eggs finished: 15
Eggs to go: 165
Eggs total: 180